+++ src/scenes/search/advanced/Memoire.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Row, Col, Container } from "reactstrap";
-import { Link } from "react-router-dom";
 import { ReactiveBase, ReactiveList } from "@appbaseio/reactivesearch/lib";
 import ExportComponent from "../components/export";
 import QueryBuilder from "../components/QueryBuilder";
 import Memoire from "../../../entities/Memoire";
-import { es_url, bucket_url } from "../../../config.js";
+import { es_url } from "../../../config.js";
 import Header from "../components/Header";
+import Card from "../components/MemoireCard";
 
 export default class Search extends React.Component {
   constructor(props) {
@@ -85,41 +85,3 @@ export default class Search extends React.Component {
     );
   }
 }
-
-const Card = ({ data }) => {
-  let image = "";
-  if (data.IMG.indexOf("memoire") === 0) {
-    image = `${bucket_url}${data.IMG}`;
-  } else if (data.IMG) {
-    image = `${data.IMG}`;
-  } else {
-    image = require("../../../assets/noimage.jpg");
-  }
-
-  return (
-    <Link
-      style={{ textDecoration: "none" }}
-      to={`/notice/memoire/${data.REF}`}
-      className="card"
-      key={data.REF}
-    >
-      <img src={image} alt="Lien cassé" />
-      <div className="content">
-        <div style={{ display: "flex" }}>
-          <h2>{data.TICO}</h2>
-          <span>{data.REF}</span>
-        </div>
-        <div>
-          <p>{data.LOCA}</p>
-          <p>{data.EDIF}</p>
-          <p>{data.LEG}</p>
-          <p>{data.OBJT}</p>
-          <p>{data.DATPV}</p>
-          <p>{data.AUTP}</p>
-          <p>{data.SERIE}</p>
-          <p>{data.TITRE}</p>
-        </div>
-      </div>
-    </Link>
-  );
-};
