+++ src/scenes/search/advanced/Merimee.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { Row, Col, Container } from "reactstrap";
-import { Link } from "react-router-dom";
 import { ReactiveBase, ReactiveList } from "@appbaseio/reactivesearch";
 import ExportComponent from "../components/export";
 import Merimee from "../../../entities/Merimee";
 import QueryBuilder from "../components/QueryBuilder";
-import { es_url, bucket_url } from "../../../config.js";
+import { es_url } from "../../../config.js";
 import Header from "../components/Header";
+import Card from "../components/MerimeeCard";
 
 export default class Search extends React.Component {
   constructor(props) {
@@ -86,39 +86,3 @@ export default class Search extends React.Component {
     );
   }
 }
-
-function getMemoireImage(memoire) {
-  if (!memoire.length || !memoire[0].url) {
-    return require("../../../assets/noimage.jpg");
-  }
-  let image = memoire[0].url;
-
-  image = image.indexOf("www") === -1 ? `${bucket_url}${image}` : image;
-  return image;
-}
-
-const Card = ({ data }) => {
-  let image = getMemoireImage(data.MEMOIRE);
-  return (
-    <Link
-      style={{ textDecoration: "none" }}
-      to={`/notice/merimee/${data.REF}`}
-      className="card"
-      key={data.REF}
-    >
-      <img src={image} alt="Lien cassé" />
-      <div className="content">
-        <div style={{ display: "flex" }}>
-          <h2>{data.TICO}</h2>
-          <span>{data.REF}</span>
-        </div>
-        <div>
-          <p>{data.DOMN}</p>
-          <p>{data.DENO.join(", ")}</p>
-          <p>{data.LOCA}</p>
-          <p>{data.AUTR.join(", ")}</p>
-        </div>
-      </div>
-    </Link>
-  );
-};
