+++ src/scenes/search/simple/Memoire.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { Row, Col, Container } from "reactstrap";
-import { Link } from "react-router-dom";
 import {
   ReactiveBase,
   DataSearch,
@@ -9,8 +8,9 @@ import {
 } from "@appbaseio/reactivesearch/lib";
 import { MultiList } from "pop-shared";
 import ExportComponent from "../components/export";
-import { es_url, bucket_url } from "../../../config.js";
+import { es_url } from "../../../config.js";
 import Header from "../components/Header";
+import Card from "../components/MemoireCard";
 
 const FILTER = [
   "mainSearch",
@@ -161,41 +161,3 @@ export default class Search extends React.Component {
     );
   }
 }
-
-const Card = ({ data }) => {
-  let image = "";
-  if (data.IMG.indexOf("memoire") === 0) {
-    image = `${bucket_url}${data.IMG}`;
-  } else if (data.IMG) {
-    image = `${data.IMG}`;
-  } else {
-    image = require("../../../assets/noimage.jpg");
-  }
-
-  return (
-    <Link
-      style={{ textDecoration: "none" }}
-      to={`/notice/memoire/${data.REF}`}
-      className="card"
-      key={data.REF}
-    >
-      <img src={image} alt="Lien cassé" />
-      <div className="content">
-        <div style={{ display: "flex" }}>
-          <h2>{data.TICO}</h2>
-          <span>{data.REF}</span>
-        </div>
-        <div>
-          <p>{data.LOCA}</p>
-          <p>{data.EDIF}</p>
-          <p>{data.LEG}</p>
-          <p>{data.OBJT}</p>
-          <p>{data.DATPV}</p>
-          <p>{data.AUTP}</p>
-          <p>{data.SERIE}</p>
-          <p>{data.TITRE}</p>
-        </div>
-      </div>
-    </Link>
-  );
-};
