+++ src/scenes/search/simple/Palissy.js
@@ -11,6 +11,7 @@ import ExportComponent from "../components/export";
 import { MultiList } from "pop-shared";
 import { es_url, bucket_url } from "../../../config.js";
 import Header from "../components/Header";
+import Card from "../components/PalissyCard";
 
 const FILTER = [
   "mainSearch",
@@ -178,70 +179,3 @@ export default class Search extends React.Component {
     );
   }
 }
-
-function getMemoireImage(memoire) {
-  if (!memoire.length || !memoire[0].url) {
-    return require("../../../assets/noimage.jpg");
-  }
-  let image = memoire[0].url;
-  image = image.indexOf("www") === -1 ? `${bucket_url}${image}` : image;
-  return image;
-}
-
-const Card = ({ data }) => {
-  const image = getMemoireImage(data.MEMOIRE);
-
-  if (data.REF.startsWith("PA") || data.REF.startsWith("PM")) {
-    return (
-      <Link
-        style={{ textDecoration: "none" }}
-        to={`/notice/palissy/${data.REF}`}
-        className="card"
-        key={data.REF}
-      >
-        <img src={image} alt="Lien cassé" />
-        <div className="content">
-          <div style={{ display: "flex" }}>
-            <h2>{data.TICO}</h2>
-            <span>{data.REF}</span>
-          </div>
-          <div>
-            <p>{data.LOCA}</p>
-            <p>{data.EDIF}</p>
-            <p>{data.AUTR.join(" ; ")}</p>
-            <p>{data.CATE.join(" ; ")}</p>
-            <p>{data.MATR.join(" ; ")}</p>
-            <p>{data.SCLE.join(" ; ")}</p>
-            <p>{data.DEPL}</p>
-            <p>{data.STAT.join(" ; ")}</p>
-            <p>{data.DPRO}</p>
-            <p>{data.DOMN}</p>
-          </div>
-        </div>
-      </Link>
-    );
-  }
-
-  return (
-    <Link
-      style={{ textDecoration: "none" }}
-      to={`/notice/palissy/${data.REF}`}
-      className="card"
-      key={data.REF}
-    >
-      <img src={image} alt="Lien cassé" />
-      <div className="content">
-        <div style={{ display: "flex" }}>
-          <h2>{data.TICO}</h2>
-          <span>{data.REF}</span>
-        </div>
-        <div>
-          <p>{data.DOMN}</p>
-          <p>{data.DENO.join(" ; ")}</p>
-          <p>{data.LOCA}</p>
-          <p>{data.AUTR.join(" ; ")}</p>
-        </div>
-      </div>
-    </Link>
-  );
-};
